Add mutt spk
Configured for reading Maildir mail from Synology Mail Server (if installed and user accounts are enabled for login). Reading mail from remote folders works as well although I decide not to make it configurable in wizard as it seems safer not to touch user configs in homes nor put email credentials in one global config.
This is the latest development release and not the last stable which is 7 years old. Even Red Hat uses 1.5.20 so it should be fine.
Closes #723.