Improve dsm-control.sh script.
Mount ${CHROOTTARGET}/tmp to be tmpfs, so that /tmp is created at start the chroot and cleared up when stopping the chroot, as per http://wiki.gentoo.org/wiki/Chroot.
Added warning to dsm-control.sh script for chroot if not started.
Fix for /dev/fd errors in emerge.
Unmount all mount points in the chroot env with a for loop. Some filesystems are mounted automatically by services, such as sysfs (which mounts ${CHROOTTARGET}/sys/fs/fuse/connections) and this stops dsm-control.sh from being able to unmount all filesystems when the chroot is stopped.
Add restart functionality to control script.
Add ability to mount custom directories in chroot.
Add log case to dsm-control.sh to allow the display of the log file /var/log/messages from inside the chroot in DSM Package Center.
Print message to user if dsm-control.sh is called with no arguments supplied.