Respect home directory for (existing) Radarr user
When creating a Radarr user (`radarr`), prior to installation of the package it should respect home directory.
Upon starting the service it will create the __.pid__ in specified home directory, while __start-stop-status__ script only checks for __.pid__ file in installation directory.
Radarr does not start when user home service is not enabled, so there is no need for fallback to installation directory in case home directory does not exist.