Extend Generic service set_syno_permissions to add all permissions
This way it can be used in packages to grant the correct permissions after updates from DSM5 or old-style packages to DSM6.
set_syno_permissions  handles all permissions so no need to use synoshare for that anymore when GROUP is set.
Permissions were extended to also allow changing of permissions on folders/files by group-members. This was also not possible with just synoshare.