feat(vim): Upgrade to vim 8.2 and build with huge features + dynamic python3
Configuring `vim` with dynamic Python 3 support means it will attempt to
load `libpython` using `dlopen`, allowing it to gracefully fail if it is
unavailable. This removes the need for a dependency on the `python3`
SPK and avoids shipping of a redundent set of Python 3 bins/libs within
the `vim` SPK package.