Add libicu (International Components for Unicode) version 67
Preparatory commit for package (znc) which would like to use this, but
the build environment lacks support for libicu thus far. My DS418play
comes with version 56.1, but I was unable to build that version due
to compilation errors. We probably don't want to either; Unicode itself
has progressed from Unicode 8.0 to Unicode 13.0 in the meanwhile.

In order to cross-compile, libicu requires a native build. The process
was inspired by its documentation describing the way to cross-compile.

While this works for me, I am not sure about cross-compiling for other
platforms. My understanding of both spksrc, libicu and build prcesses
in general are very lacking.