Update ZNC to 1.8.1 including ICU / Python3 support.
Including Python by building modpython.so dragged in a dependency on ICU
(aka libicu), but because this also enables some configurable features
for ZNC, that is not something negative.

It adds a SPK dependency on the Python3 package. Perhaps one can remove
this since modpython is completely optional, but I don't understand how
everything work together well enough to be confident enough it won't
break anything.

Note that this still uses the legacy ./configure that has been removed
from upstream after this versions release in favor of a CMake solution.
(But for this version it is still perfectly present & functional.)

A future update will have to tackle the change in build systems, but at
present that is beyond my ability. (It tries to disguise itself as a GNU
configure script, but that messes with spksrc cross-compiling because
of flags that aren't supported. Turning GNU_CONFIGURE off does not fix
it, nor does switching the include to CMake, which marks the end of this
novices patience with this ordeal.)