syncthing: Do not cache .stignore file patterns by default.
Change the default for the cacheIgnoredFiles option to false.  From
the Syncthing documentation [1]:

    Whether to cache the results of ignore pattern
    evaluation. Performance at the price of memory. Defaults to false
    as the cost for evaluating ignores is usually not significant.

The DiskStations are usually rather tight on RAM, so deviating from
Syncthing's default here is probably a bad idea for most users.  It
only matters when many (or very complex) ignore patterns are used in a
folder, so usually for power users.  Those will certainly be able to
change the option manually.

Note that this also only applies for new installations, because
config.xml is (and must be) kept during updates.

[1]: https://docs.syncthing.net/users/config.html