syncthing: Upgrade to v1.15.1 and ignore folder permissions by default. (#4527)
* syncthing: Incorporate upstream release v1.14.0.

Only package version and hashes updated.  Built successfully for
all-supported.

Note that the previous package version should have already
self-updated an installed Syncthing binary when the service was
running long enough.

* syncthing: Ignore permissions by default for new shared folders.

Utilize the newly introduced "folder default configuration" [1]
feature to preset the ignorePerms option to true.  This will not touch
any existing folders, but set the option when configuring a new folder
to be synced.

Specifically on DSM, the ignorePerms option should definitely be used
with every Syncthing folder.  Otherwise, it will possibly try to
override the UNIX permissions on synced files, which will reset the
ACL-based permissions on the content.  Possibly making them unusable
for the rest of the DSM services.

Note that this also only applies for new installations, because
config.xml is (and must be) kept during updates.

[1]: https://github.com/syncthing/syncthing/pull/7131

* syncthing: Do not cache .stignore file patterns by default.

Change the default for the cacheIgnoredFiles option to false.  From
the Syncthing documentation [1]:

    Whether to cache the results of ignore pattern
    evaluation. Performance at the price of memory. Defaults to false
    as the cost for evaluating ignores is usually not significant.

The DiskStations are usually rather tight on RAM, so deviating from
Syncthing's default here is probably a bad idea for most users.  It
only matters when many (or very complex) ignore patterns are used in a
folder, so usually for power users.  Those will certainly be able to
change the option manually.

Note that this also only applies for new installations, because
config.xml is (and must be) kept during updates.

[1]: https://docs.syncthing.net/users/config.html

* syncthing: Remove / update obsolete default options.

Drop the symlinksEnabled and weakHashSelectionMethod entries from the
default config file template.  See [1] and [2].

Switch from the minHomeDiskFreePct option to the new syntax
minHomeDiskFree unit="%".

Note that this also only applies for new installations, because
config.xml is (and must be) kept during updates.  Syncthing migrates
old configuration formats on its own.

[1]: https://github.com/syncthing/docs/commit/392dbd0b
[2]: https://github.com/syncthing/syncthing/pull/4767

* syncthing: Add myself (@acolomb) as maintainer.

* Syncthing: Drop busybox and DSM 5 support

Required to prepare for DSM 7.0 support

* Syncthing: fixes

* Framework: log the actual command after service_prestart()

* syncthing: Incorporate upstream release v1.15.1.

Only package version and hashes updated.  Built successfully for
all-supported.

Note that the previous package version should have already
self-updated an installed Syncthing binary when the service was
running long enough.

Co-authored-by: Sebastian Schmidt <publicarray@users.noreply.github.com>