cmake+ninja: Allow using CMAKE_USE_DESTDIR from ninja
The CMAKE_USE_DESTDIR is by default enabled (i.e. set to 1)
and doing so adds to the environment DESTDIR=<install_path>
to the installation call. Some dependencies will not install
properly when DESTDIR is set (i.e. umurmur and intel-media-driver).
As such CMAKE_USE_DESTDIR must be set to 0 so the DESTDIR variable
is not being set in the environment when installing.

This functionality is now added to ninja to allow not setting
DESTDIR when needed.