spk.mk: Fix pre-build-native for multi-arch specific builds
When invoking make -j2 arch-armada38x it tries to build
both arch-arch-armada38x-6.1 and arch-arch-armada38x-7.0.
This fails when hitting native/* dependencies as they must
be built in sequence first, and only then switch to parallel
for all arch-* requested.

This issue was faced originally with make all-supported and
fixed by calling first pre-build-native as needed.  The change
here addresses this to allow the same building process when
multiple arch-* are passed by the user.