framework: Add support for the SSL certificate resource config.
Introduce new variables SERVICE_CERT and SERVICE_CERT_RELOAD for use
in spk Makefiles.  If SERVICE_CERT is defined, an appropriate resource
configuration "certificate-config" is generated for the named service,
which must also be defined in the $(SPK_NAME).sc file.  The
display_name_i18n field seems to be required, so set it to the same as
the display_name for now.

Only one service per package is supported for now, as more would
require more complex syntax per package.  Similar to how there is a
single SERVICE_PORT.

A reloader script should be specified in the SERVICE_CERT_RELOAD
variable, as a package relative path.  It is copied to the resource
file verbatim and executed by DSM when the certificate is updated or
replaced.  Installation of this script must be taken care of by the
package-specific Makefile.