syncthing: Avoid overwriting newer binary from internal updater (fixes #5142) (#5181)
* syncthing: Define shortcut variable for binary location.
* syncthing: Simplify HOME setting in service-setup.
The variable is needed for any function that tries to call the
syncthing binary.  So define and export it once at the top instead of
in every function that needs it.
* syncthing: Retain installed binary if newer than the packaged one.
Check the version strings from the installed and packaged binaries
respectively during the preupgrade step.  Keep a backup in the
SYNOPKG_TEMP_UPGRADE_FOLDER if necessary and restore that during the
postupgrade step.
* Incorporate latest upstream release v1.20.0.
* syncthing: Bump dependency to Go 1.18 and remove native/go_1.17.