Default "ignorePerms" option to true for new folders.
Include a minimal folder defaults configuration in the shipped
config.xml file, which enables the option.  When Syncthing tries to
apply POSIX permissions on DSM, that will clear out the existing ACLs,
leading to strange permission problems.

This will only apply to new installations, as the config template is
only used during the first installation.