Upgrade sonarr build to latest
Motivation: The recent 7.1 DSM upgrade kicked Sonarr into a bad state
and required a repair.
However, there are datastore changes between the current version and the
latest such that the data is incompatible with the current SPK version,
meaning that Sonarr would not properly load and as a result also can't
run the upgrade from the internal upgrader
This left me with the choices of either restarting from scratch (no good
backups), or compiling a package from source.
Given I did the later, it seemed sensible to push a PR for everyone
else.

Checklist:

[x] Built Sonarr package successfully (since changes are limited to it)
[x] Package upgrade completed successfully
[x] New installation of package completed successfully