[firebird] Move config files to $SYNOPKG_PKGVAR
Declare the FB_* paths in the C(XX)FLAGS in the cross/Makefile instead
of in the cross_config include files. These paths affect the runtime
behavior of Firebird and must match the path on the installed system.

The paths for configuration ($SYNOPKG_PKGVAR) however differ during
install and runtime, as the installer script moves the content of
$PREFIX/var to $SYNOPKG_PKGVAR during installation.

The change of FB_CONFDIR also changes the PREFIX a.k.a. root directory
of the whole Firebird installation and when modified, all other paths
must be explicitly specified.