Add go2rtc package
This commit adds the go2rtc package, a camera streaming application with support for various protocols. The package version is 1.7.1. The Makefile, PLIST, and digests for the package are created. The package is also added to the Synology package manager with the necessary configuration files, service setup script, and user interface files for installation and upgrade wizards. The package is configured to run as a background service with a dedicated user group.