Add go2rtc package (#5889)
* Add go2rtc package
This commit adds the go2rtc package, a camera streaming application with support for various protocols. The package version is 1.7.1. The Makefile, PLIST, and digests for the package are created. The package is also added to the Synology package manager with the necessary configuration files, service setup script, and user interface files for installation and upgrade wizards. The package is configured to run as a background service with a dedicated user group.

* Update Maintainer in go2rtc Makefile
The maintainer of the go2rtc Makefile has been updated from AlexxIT to skrashevich. This change reflects the current responsibility for the maintenance of this file.

* fix PKG_DIST_{NAME,SITE,FILE} vars in cross/go2rtc/Makefile
Co-authored-by: hgy59 <hpgy59@gmail.com>

* Update go2rtc.png

* Add new configuration for go2rtc-srtp in go2rtc.sc and update service-setup.sh to include the new configuration if the package status is INSTALL

* backport https://github.com/AlexxIT/go2rtc/pull/652

* fix srtp port typo

* Update cross/go2rtc/Makefile
Co-authored-by: hgy59 <hpgy59@gmail.com>

* update go2rtc configuration
- install go2rtc.yaml file
- remove environment-variables file
- remove obsolete variables
  - DSM_UI_DIR: no app icon is created in DSM UI (and app is the default DSM_UI_DIR anyway)
  - SPK_COMMANDS: there is no cli tool installed (go2rtc is used in SERVICE_COMMAND only)
- use sed to inject wizard variables into config file at package installation
- remove upgrade wizard

* update config
- avoid color codes in log output
- make config file more expressive by adding all listening ports (even the default ones)
- update WebRTC protocol description
- remove obsolete PATCH_LEVEL and include in cross/go2rtc/Makefile

* force use of ipv4 for srtp

* v1.8.1

---------

Co-authored-by: hgy59 <hpgy59@gmail.com>