Update go2rtc version to 1.8.4
This commit updates the go2rtc package version from 1.8.2 to 1.8.4. The changes include updates to the Makefile and digests files. The changelog has also been updated to reflect the new features and bug fixes introduced in this version. These include improved support for incoming MJPEG, fixes for https source in some cases, and the addition of new GoPro source for streaming from GoPro cameras among others.