feat(go2rtc): add patch to refactor homekit helpers and introduce `maxByteSlice`
The commit introduces a new patch file for the go2rtc project which refactors functions in the `homekit/helpers.go` file. This patch replaces the usage of the external `slices` package with a newly added function `maxByteSlice` within the same file, streamlining the extraction of maximum profileID and level values from byte slices. The `PATCHES_LEVEL` variable is also added to the Makefile, likely indicating the system should apply one level of patches during build.