Enhance Docker platform specification in documentation
The current documentation lacks clarification on Docker platform specifications,
leading to errors when attempting to run the application on non-linux/amd64 machines.

       docker: no matching manifest for linux/arm64/v8 in the manifest list entries.

To ensure seamless execution on linux/arm architectures,
specify the appropriate platform in the Docker command line.
This adjustment aims to enhance the user experience and prevent potential issues.