intel-opencl-clang: Enable shared libraries
It hapens that later-on IGC builds-up a compile-time tool called
CMCLTranslatorTool which isn't built for 'NATIVE' and rather
requires enforcing loading libLLVMGenXIntrinsics.so from
intel-vc-intrinsincs.  In turn, when built "static" (default mode),
libLLVMGenXIntrinsics.so ends-up containing llvm libraries as well,
which in turns CMCLTranslatorTool also contain, leading to a duplicate
existance of similar calls, thus segfaulting core-dump.

Note that loading build resulting libraries for this CMCLTranslatorTool
really is an ugly hack that can only work for x86_64 built on x86_64
with compatible C libraries.