Ente: Initial package add for the Photo Manager
Ente is an encrypted self-hosted photo management app.

This packages it natively.
It depends on Minio object storage and Postgres.

Minio is already a syno package, we can use that.
Synology NASs come with Postgres, so the readme has instructions on how
to just use that, but you could connect it to any postgres you wanted
to.

Fixes #6305.

Relates to https://github.com/ente-io/ente/discussions/4057.